[Core] Add support for multi-dimensional array writing.
The problem with the old code is that it assumed that any Array was uni-dimensional. Therefore, it would end up with an invalid cast exception.
When writing any array into the binary message, flatten the array. Thus, when reading the array back, we read it as uni-dimensional. It should not impact anything.

There is a fast-path for rank1 arrays, as the foreach is rewritten to handle generalized iteration.

Bug 59805 - Visual Studio for Mac potential issue on unit testing support